Change Into to long

+++ ReviewApp.IServices/IAuthService.cs
@@ -10,7 +10,7 @@ namespace ReviewApp.IServices
 {
     public  interface IAuthService
     {
-        Task<(string AccessToken, string RefreshToken)> GenerateJwtToken(string EmailAddress, int UserId);
+        Task<(string AccessToken, string RefreshToken)> GenerateJwtToken(string EmailAddress, long UserId);
         Task<(string AccessToken, string RefreshToken, User user)> RefreshAccessTokenAsync(string refreshToken);
     }
 }
